move common to common.py
